Replace non-allowed characters from rowkeys in Table Storage repositories
