+++ queue/src/fiskaltrust.Middleware.Storage.AzureTableStorage/Repositories/DE/AzureTableStorageFailedFinishTransactionRepository.cs
@@ -3,9 +3,8 @@ using System.Threading.Tasks;
 using Azure.Data.Tables;
 using fiskaltrust.Middleware.Contracts.Data;
 using fiskaltrust.Middleware.Contracts.Models.Transactions;
+using fiskaltrust.Middleware.Storage.AzureTableStorage.Helpers;
 using fiskaltrust.Middleware.Storage.AzureTableStorage.Mapping;
-using fiskaltrust.Middleware.Storage.AzureTableStorage.TableEntities;
-using fiskaltrust.storage.V0;
 
 namespace fiskaltrust.Middleware.Storage.AzureTableStorage.Repositories.DE
 {
@@ -18,11 +17,11 @@ namespace fiskaltrust.Middleware.Storage.AzureTableStorage.Repositories.DE
 
         public async Task InsertOrUpdateTransactionAsync(FailedFinishTransaction transaction) => await InsertOrUpdateAsync(transaction).ConfigureAwait(false);
 
-        public async Task<bool> ExistsAsync(string cbReceiptReference) => await GetAsync(cbReceiptReference).ConfigureAwait(false) != null;
+        public async Task<bool> ExistsAsync(string cbReceiptReference) => await GetAsync(ConversionHelper.ReplaceNonAllowedKeyCharacters(cbReceiptReference)).ConfigureAwait(false) != null;
 
         protected override void EntityUpdated(FailedFinishTransaction entity) { }
 
-        protected override string GetIdForEntity(FailedFinishTransaction entity) => entity.cbReceiptReference;
+        protected override string GetIdForEntity(FailedFinishTransaction entity) => ConversionHelper.ReplaceNonAllowedKeyCharacters(entity.cbReceiptReference);
 
         public async Task InsertOrUpdateAsync(FailedFinishTransaction storageEntity)
         {
@@ -38,7 +37,7 @@ namespace fiskaltrust.Middleware.Storage.AzureTableStorage.Repositories.DE
                 return null;
             }
 
-            var entity = new TableEntity(Mapper.GetHashString(src.FinishMoment.Ticks), src.cbReceiptReference)
+            var entity = new TableEntity(Mapper.GetHashString(src.FinishMoment.Ticks), GetIdForEntity(src))
             {
                 { nameof(FailedFinishTransaction.cbReceiptReference), src.cbReceiptReference },
                 { nameof(FailedFinishTransaction.CashBoxIdentification), src.CashBoxIdentification },
